update
- Jwt token 확인해주는 proxyServer

- AccessToken 다시 발급 받기 위한 CreateNewAccessToken

- Header로 refreshToken 받기

- User 객체로 accessToken 넘겨주기

- countAlarm 같은 count 객체를 반환 받는 메서드는
count = -1 = invalidate
count = -2 = expired

- delete부분 모든 기능 nickname 추가

to do

- 로그 아웃 기능 추가
-> 해당 nickname, accessToken 받아서 만료 시키기

- 기존에 있던 token 다 지우기

- User에 refresh, access Token 추가하기

- isvalidate (메인 페이지 들어가자마자 토큰 유효성 검사)
-> 유효하면, 로그인 된 창 표시
-> 유효하지 않다면, 로그 아웃 된 창 표시

- 용어집 업데이트 (delete 부분)
-> Question
-> Answer
-> Comment
-> Alarm
-> createNewAccessToken

+++ dev/back-end/proxy/src/main/java/reverseproxy/proxy/GraphQLMainServer/Delete/DeleteAlarm.java
@@ -1,13 +1,23 @@
 package reverseproxy.proxy.GraphQLMainServer.Delete;
 
 import com.google.gson.Gson;
+import graphql.schema.DataFetchingEnvironment;
 import org.springframework.stereotype.Component;
 import reverseproxy.proxy.Entity.Alarm;
 import reverseproxy.proxy.GraphQLMainServer.ConnectMainServer;
 
 @Component
 public class DeleteAlarm extends ConnectMainServer {
-    public Alarm deleteAlarm(String alarm_id) {
+    public Alarm deleteAlarm(String alarm_id, String nickname, DataFetchingEnvironment env) throws Exception {
+        String authorized = checkJwt(nickname, env);
+        String invalidate = "invalidate";
+        String expired = "expired";
+        if (authorized.equals(invalidate)) {
+            return new Alarm(invalidate);
+        }
+        if (authorized.equals(expired))
+            return new Alarm(expired);
+
         //region Query
         String query = "mutation{\n" +
                 "    deleteAlarm(alarm_id \": " + alarm_id + "\"){\n" +
